Divides class into header and .cpp files
